feat: restructuring application files
