+++ android/app/src/main/java/com/example/app/MainActivity.java
@@ -2,7 +2,7 @@ package com.example.app;
 
 import android.os.Bundle;
 
-import com.example.app.plugins.subscriptions.AppSubscriptionPlugin;
+import com.example.app.presentation.subscriptions.AppSubscription;
 import com.getcapacitor.BridgeActivity;
 
 public class MainActivity extends BridgeActivity {
@@ -12,6 +12,6 @@ public class MainActivity extends BridgeActivity {
         super.onCreate(savedInstanceState);
         registerPlugin(EchoPlugin.class);
         registerPlugin(HelloPlugin.class);
-        registerPlugin(AppSubscriptionPlugin.class);
+        registerPlugin(AppSubscription.class);
     }
 }
